requirements.txt
streamlit>=1.40.0
pandas
openpyxl
pymupdf
pillow